Firmware changes to support scanning sequence and fix packet bugs

+++ firmware/lib/utility/timing.h
@@ -1,4 +1,6 @@
 #pragma once
 
+#include <stdint.h>
+
 void delay_ms(unsigned int ms);
-long long unsigned get_tick_ms();
+uint32_t get_tick_ms();
